added in a config module, added in a util_funcs module. fixed function call in experiments, still needs entitiy data for them to run correctly
