+++ config/__init__.py
@@ -1,0 +1,1 @@
+from .model_config import MODEL_CONFIG
