+++ config/model_config.py
@@ -1,0 +1,6 @@
+# Baseline model configuration
+MODEL_CONFIG = {
+    'n_embd': 128,
+    'n_head': 4,
+    'n_layer': 4,
+}
