+++ util_funcs/__init__.py
@@ -1,0 +1,1 @@
+from .seq_util import find_max_sequence_length
